refactor(#5031): W3 extract session_idle_cleanup from discord/mod.rs (move-only)
